\name{priorHistone_multi}
\alias{priorHistone_multi}
\title{Process Histone data to cluster positions.}
\usage{
priorHistone_multi(object, dnaseIndex, outfileLoc="./", outfile = "histoneOnly")
}
\arguments{
\item{object}{A "Prior" S4 object generated by 
\code{priorHistone_init}. If not provided, the default value will be NULL.}

\item{dnaseIndex}{Index of data which will be used as DNase-seq data in the
model. Default value will be 1.}

\item{outfileLoc}{Directory to store processed files. Default is set to "./"}

\item{outfile}{Infix of outfile name. Default is set to "histoneOnly" to
indicate the prior is built by Histone datasets only.}
}
\value{
A new "Prior" object is created with new information updated:

\item{dnaseName}{Name of DNase-seq dataset is the selected
  histone ChIP-seq dataset used as DNase-seq in the prior construction.}

\item{dnaseKnots}{DNase knots (from histone ChIP-seq file used as DNase-seq)
  for B-spline functions. They are the 90, 99 and 99.9th percentiles of read counts.}


  \item{dnaseThres}{A vector of DNase-seq group created to generate aggregated ChIP
data. After alignment, positions which have the same DNase-seq read count are clustered into
  one group. \code{dnaseThres} is the corresponding read count number
  in each group. Each count is corresponding to one group and the grouping data 
partitions the whole genome into multiple segments.}


\item{posLoc_bychr}{Location of the files containing the group index of
  each segment of the genome.}

\item{dnaseAlign}{Aligment information of the DNase file (from histone
  ChIP-seq file used as DNase-seq) from bowtie.}

\item{dataNum}{Number of data sets ("DNase-seq" and histone ChIP-seq data sets).}
\item{histoneNum}{Number of histone ChIP-seq files.}
\item{histoneName}{Name of histone ChIP-seq files.}
\item{histoneAlign}{Aligment information of the histone ChIP-seq file(s) from bowtie.}

}
\description{
Construct required files for generating priors for ChIP-seq data based on only Histone data.
}
\details{
This function processes the Histone files and generates module for further
analysis.
}
\examples{
\dontrun{
priorobject = priorHistone_multi(object, dnaseIndex,
outfileLoc = "./", outfile)

}
}
\author{
Xin Zeng,  M. Constanza Rojo-Alfaro, Ye Zheng
}
\keyword{methods}
\keyword{models}

