\name{readAllocate}
\alias{readAllocate}
\title{Allocate ChIP-seq reads}
\usage{
readAllocate(object, outfileLoc = "./", outputFormat = NULL,
ChipThres = NULL, chipFile = NULL, bowtieIndex = NULL, csemDir = NULL,
bowtieDir = NULL, vBowtie = NULL, mBowtie = NULL, pBowtie = NULL,
fragL = NULL)
}
\arguments{
\item{object}{A "Prior" S4 object generated by \code{priorGenerate}. If
  object=NULL, then extra parameters, which are chipFile, bowtieIndex, csemDir,
  bowtieDir, vBowtie, mBowtie, pBowtie, fragL, need to be defined and CSEM will allocate-multi reads without using prior information (see details for more information).}

\item{outfileLoc}{Directory to store processed files.}

\item{outputFormat}{Output Format. Default is NULL and will only return
  results in bam format. Possible values are tagAlign and bed}

\item{ChipThres}{For thresholding reads. It will select reads with  scores higher than\cr
  \code{ChipThres} (allocation probability*1000). Default set at NULL and will not threshold reads.}

\item{chipFile}{Default set as NULL, only needed when object=NULL. ChIP-Seq files, in fastq format.}

\item{bowtieIndex}{Default set as NULL, only needed when object=NULL. Bowtie index, used in bowtie aligning}

\item{csemDir}{Default set as NULL, only needed when object=NULL. Directory where CSEM was installed.}

\item{bowtieDir}{Default set as NULL, only needed when object=NULL. Directory where Bowtie was installed}

\item{vBowtie}{Default set as NULL, only needed when object=NULL. Bowtie parameter. In -v mode, alignments may have
no more than vBowtie mismatches, where \code{v} may be a number from 0
through 3 set using the -v option. Default value is 2.}

\item{mBowtie}{Default set as NULL, only needed when object=NULL. Bowtie parameter. -m parameter instructs bowtie to
refrain from reporting any alignments for reads having more than
\code{mBowtie} reportable alignments. Default value is 99 allowing
multi-reads alignment.}

\item{pBowtie}{Default set as NULL, only needed when object=NULL. The -p option causes Bowtie to launch a specified
number of parallel search threads. Each thread runs on a different
processor/core and all threads find alignments in parallel. Default value is 8.}

\item{fragL}{Default set as NULL, only needed when object=NULL. Fragment length.}

}
\value{
An updated "Prior" object is returned together with the aligned ChIP-seq
output files.
\item{chipSAM}{Location of aligned ChIP-seq files in SAM format.}
\item{chipAllocate}{Location of allocated ChIP-seq files in bam format.}
\item{chipFormat}{Location of allocated ChIP-seq files in other format (tagAlign or bed).}
\item{outfileLoc}{Directory where processed files are (given as an argument).}

Or if no prior is built and allocate ChIP-seq by CSEM directly, a new
"Prior" object will be created.
\item{chipName}{Name of ChIP-seq dataset(s).}
\item{chipNum}{Number of ChIP-seq dataset(s).}
\item{chipAlign}{ChIP-seq alignment summary feedback from bowtie.}
\item{chipSAM}{Location of aligned ChIP-seq in SAM format.}
\item{chipAllocate}{Location of the final allocated ChIP-seq output
  by Permseq in BAM format.}
\item{chipFormat}{Location of the final allocated ChIP-seq
  output by Permseqin other selected format (tagAlign or bed).}
\item{fragL}{Fragment length.}
\item{bowtieInfo}{List of bowtie information used: bowtieIndex, bowtieDir, vBowtie, mBowtie and pBowtie.}
\item{csemDir}{Directory where CSEM was installed.}
\item{outfileLoc}{Directory to store output files.}


}
\description{
Allocate ChIP-Seq multi-reads using prior information.
}
\details{
By default, it will allocate multi-reads using prior information from
the object generated by \cr
\code{priorGenerate}.

If prior information is not available (DNase and/or histone), do not run
\code{priorProcess} and \code{priorGenerate}. Instead,  set
\code{object} as NULL and give the extra parameters: chipFile,
bowtieIndex, csemDir, bowtieDir, vBowtie, mBowtie, pBowtie and
fragL. CSEM will allocate the ChIP-seq.

\code{plot()}, \code{summary()}, \code{names()} and \code{print()} methods included. To obtain the alignment information from bowtie use \code{summary()}.
}




\examples{
\dontrun{
readAllocate(object, outfileLoc = "./", outputFormat = "tagAligne")
}
}


\author{
Xin Zeng, M. Constanza Rojo-Alfaro, Ye Zheng.
}

